Reorganize CLI utility helpers
