+++ src/valkyrie/cli/benchmark/tasks.py
@@ -3,8 +3,8 @@ from pathlib import Path
 import click
 
 from valkyrie.cli.exceptions import TrackerServiceError
+from valkyrie.cli.health import check_tracker_service_health
 from valkyrie.cli.tracker_service import TrackerService
-from valkyrie.cli.utils import check_tracker_service_health
 
 
 @click.command(
