+++ src/valkyrie/cli/run/fetch.py
@@ -3,8 +3,9 @@ from uuid import UUID
 import click
 
 from valkyrie.cli.exceptions import TrackerServiceError
+from valkyrie.cli.health import check_tracker_service_health
+from valkyrie.cli.run.status import format_benchmark_status, stream_benchmark_status
 from valkyrie.cli.tracker_service import TrackerService
-from valkyrie.cli.utils import check_tracker_service_health, format_benchmark_status, stream_benchmark_status
 
 
 @click.command(
