+++ src/valkyrie/cli/run/start.py
@@ -7,18 +7,135 @@ from tracker.types import StartBenchmarkResponse
 
 from valkyrie.cli.bundler import get_contract
 from valkyrie.cli.exceptions import BundlerError, ContractValidationError, TrackerServiceError
+from valkyrie.cli.health import check_tracker_service_health
+from valkyrie.cli.run.status import stream_benchmark_status
+from valkyrie.cli.run.task_ids import resolve_task_ids
 from valkyrie.cli.s3_client import get_contract_from_s3, push_agent
+from valkyrie.cli.table import local_time
 from valkyrie.cli.tracker_service import TrackerService
-from valkyrie.cli.utils import (
-    check_tracker_service_health,
-    format_agent_start_details,
-    format_run_start_details,
-    format_start_benchmark_response,
-    resolve_task_ids,
-    resolve_webhook_config,
-    stream_benchmark_status,
-)
 from valkyrie.schemas import AgentConfig
+
+
+COLUMN_WIDTH = 14
+
+
+def row(label: str, value: str) -> str:
+    return f"│ {label:<{COLUMN_WIDTH}} {value}"
+
+
+def cont(value: str) -> str:
+    return f"│ {' ' * COLUMN_WIDTH} {value}"
+
+
+def format_run_start_details(
+    benchmark: str, dataset: str | None, concurrency: int, slice_str: str | None, task_ids: str | None
+) -> None:
+    """Format and display the start details of a run."""
+    click.echo("┌─ Benchmark " + "─" * 67)
+    click.echo(row("Name:", benchmark))
+    click.echo(row("Dataset:", dataset or "default"))
+    click.echo(row("Concurrency:", str(concurrency)))
+    if slice_str:
+        click.echo(row("Slice:", slice_str))
+    if task_ids:
+        display = task_ids[:60] + "..." if len(task_ids) > 60 else task_ids
+        click.echo(row("Task IDs:", display))
+    else:
+        click.echo(row("Task IDs:", "all tasks"))
+    click.echo("└" + "─" * 79)
+    click.echo()
+
+
+def format_agent_start_details(
+    agent: str,
+    model: str | None,
+    secrets: tuple[tuple[str, str], ...],
+    kwargs: tuple[tuple[str, str], ...],
+    service_headers: dict[str, str],
+    webhook_secret: str | None,
+    webhook_intervals: list[int] | None,
+) -> None:
+    """Format and display the start details of an agent."""
+    click.echo("┌─ Agent " + "─" * 71)
+    click.echo(row("Agent:", agent))
+    if model:
+        click.echo(row("Model:", model))
+    if secrets:
+        for index, (env_var, secret_name) in enumerate(secrets):
+            if index == 0:
+                click.echo(row("Secrets:", f"{env_var} → {secret_name}"))
+            else:
+                click.echo(cont(f"{env_var} → {secret_name}"))
+    if kwargs:
+        click.echo(row("Kwargs:", ", ".join(f"{key}={value}" for key, value in kwargs)))
+    if service_headers:
+        click.echo(row("Headers:", ", ".join(service_headers.keys())))
+    if webhook_secret and webhook_intervals:
+        click.echo(row("Notify at:", ", ".join(f"{interval}%" for interval in webhook_intervals)))
+    click.echo("└" + "─" * 79)
+    click.echo()
+
+
+def format_start_benchmark_response(start_benchmark_response: StartBenchmarkResponse) -> None:
+    """Format and display the start run response."""
+    run_id = start_benchmark_response.benchmark_id
+
+    click.echo()
+    click.echo("┌─ Run Details " + "─" * 65)
+    click.echo(f"│ {'Benchmark:':<17} {start_benchmark_response.benchmark_name}")
+    click.echo(f"│ {'Agent:':<17} {start_benchmark_response.agent_name}")
+    click.echo(f"│ {'Run ID:':<17} {run_id}")
+    click.echo(f"│ {'Started at:':<17} {local_time(start_benchmark_response.started_at)}")
+    click.echo(f"│ {'Max concurrency:':<17} {start_benchmark_response.concurrency}")
+    click.echo(f"│ {'Total tasks:':<17} {start_benchmark_response.task_count}")
+    click.echo(f"│ {'CloudWatch:':<17} {start_benchmark_response.cloudwatch_url}")
+    click.echo(f"│ {'S3 Bucket:':<17} {start_benchmark_response.s3_bucket_url}")
+    click.echo("├" + "─" * 79)
+    click.echo(
+        f"│ {'Get results:':<17} "
+        + click.style(f"valkyrie run results {run_id} --path ./results-{run_id}.json", fg="cyan")
+    )
+    click.echo(f"│ {'Stop:':<17} " + click.style(f"valkyrie run stop {run_id}", fg="cyan"))
+    click.echo(f"│ {'Resume:':<17} " + click.style(f"valkyrie run resume {run_id}", fg="cyan"))
+    click.echo(f"│ {'Retry:':<17} " + click.style(f"valkyrie run retry {run_id}", fg="cyan"))
+    click.echo(f"│ {'Run outputs:':<17} " + click.style(f"valkyrie run outputs {run_id} --output-dir .", fg="cyan"))
+    click.echo("└" + "─" * 79)
+
+
+def validate_intervals(intervals: tuple[int, ...]) -> list[int]:
+    """Validate notification interval values."""
+    interval_list = list(intervals)
+    if len(interval_list) > 3:
+        raise click.UsageError("Maximum of 3 intervals allowed.")
+    for interval in interval_list:
+        if interval < 5 or interval > 100:
+            raise click.UsageError(f"Interval {interval} out of range. Must be between 5 and 100.")
+        if interval % 5 != 0:
+            raise click.UsageError(f"Interval {interval} must be divisible by 5.")
+    return interval_list
+
+
+def resolve_webhook_config(
+    intervals: tuple[int, ...], webhook_secret: str | None
+) -> tuple[str | None, list[int] | None]:
+    """Resolve webhook secret and intervals for a benchmark run."""
+    if intervals and not webhook_secret:
+        click.echo(
+            click.style(
+                "  Warning: --interval specified but no webhook secret configured. "
+                "Run `valkyrie config set webhook <secret-name>` first. Ignoring intervals.",
+                fg="yellow",
+            )
+        )
+        return None, None
+
+    if intervals:
+        return webhook_secret, validate_intervals(intervals)
+
+    if webhook_secret:
+        return webhook_secret, [100]
+
+    return None, None
 
 
 @click.command(
